Reorganizing methods from ClientAuthenticator into LwaService and new LwaClient; adjusted tests accordingly
