+++ src/Services/Authenticator/ClientAuthenticator.php
@@ -10,19 +10,9 @@ use GuzzleHttp\Client;
 use GuzzleHttp\ClientInterface;
 use GuzzleHttp\Handler\CurlHandler;
 use GuzzleHttp\HandlerStack;
-use Psr\SimpleCache\CacheInterface;
 
 class ClientAuthenticator implements ClientAuthenticatorInterface
 {
-    const LWA_ACCESS_TOKEN_CACHE_KEY = 'lwa_access_token';
-
-    const CACHE_LIFE_BUFFER_IN_SECONDS = 60;
-
-    /**
-     * @var CacheInterface
-     */
-    protected $cache;
-
     /**
      * @var LwaServiceInterface
      */
@@ -39,12 +29,10 @@ class ClientAuthenticator implements ClientAuthenticatorInterface
     protected $spApiConfig;
 
     public function __construct(
-        CacheInterface $cache,
         LwaServiceInterface $lwaService,
         callable $credentialProvider,
         SpApiConfig $spApiConfig
     ) {
-        $this->cache              = $cache;
         $this->lwaService         = $lwaService;
         $this->credentialProvider = $credentialProvider;
         $this->spApiConfig        = $spApiConfig;
@@ -60,12 +48,13 @@ class ClientAuthenticator implements ClientAuthenticatorInterface
      * @return ClientInterface
      * @throws LwaAccessTokenException
      */
-    public function createAuthenticatedGuzzleClient($restrictedDataToken = null)
-    {
+    public function createAuthenticatedGuzzleClient(
+        $restrictedDataToken = null
+    ) {
         if ($restrictedDataToken) {
             $accessToken = $restrictedDataToken;
         } else {
-            $accessToken = $this->rememberLwaAccessToken();
+            $accessToken = $this->lwaService->rememberLwaAccessToken();
         }
 
         $stack = $this->_buildHandlerStack();
@@ -80,41 +69,15 @@ class ClientAuthenticator implements ClientAuthenticatorInterface
         ]);
     }
 
-    /**
-     * Get the cached Login with Amazon (LWA) access token if it exists, or request a new one
-     * and save it in the cache.
-     *
-     * @return string
-     * @throws LwaAccessTokenException
-     */
-    public function rememberLwaAccessToken()
-    {
-        if ($cachedToken = $this->cache->get(self::LWA_ACCESS_TOKEN_CACHE_KEY)) {
-            return $cachedToken;
-        }
-
-        $newToken = $this->lwaService->requestNewLwaAccessToken(new Client([
-            'base_uri' => $this->spApiConfig->lwaOAuthBaseUrl,
-            'debug'    => $this->spApiConfig->debugOAuthApiCall,
-        ]));
-
-        $this->cache->set(
-            self::LWA_ACCESS_TOKEN_CACHE_KEY,
-            $newToken['access_token'],
-            $newToken['expires_in'] - self::CACHE_LIFE_BUFFER_IN_SECONDS
-        );
-
-        return $newToken['access_token'];
-    }
-
     protected function _buildHandlerStack()
     {
+        // TODO: Parameterize a HandlerStack to give developers more flexibility.
         $stack = new HandlerStack();
         $stack->setHandler(new CurlHandler());
 
         $stack->push(new AwsSignatureV4Middleware(
             $this->credentialProvider,
-            // TODO: Config-ify these values?
+            // TODO: Configify defaults & parameterize overrides for these values.
             'execute-api',
             'us-east-1'
         ));
