+++ src/Services/Lwa/LwaClientInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Glue\SpApi\OpenAPI\Services\Lwa;
+
+use Glue\SpApi\OpenAPI\Exceptions\LwaAccessTokenException;
+use GuzzleHttp\ClientInterface;
+
+interface LwaClientInterface
+{
+    /**
+     * Request a new Login with Amazon (LWA) access token.
+     *
+     * @return array
+     * @throws LwaAccessTokenException
+     */
+    public function requestNewLwaAccessToken(ClientInterface $guzzleClient);
+}
