+++ tests/Services/ClientAuthenticatorTest.php
@@ -6,18 +6,11 @@ use Aws\Credentials\CredentialsInterface;
 use Glue\SpApi\OpenAPI\Services\Authenticator\ClientAuthenticator;
 use Glue\SpApi\OpenAPI\Services\Lwa\LwaServiceInterface;
 use Glue\SpApi\OpenAPI\SpApiConfig;
-use GuzzleHttp\Client;
 use Mockery\MockInterface;
-use Psr\SimpleCache\CacheInterface;
 use Tests\TestCase;
 
 class ClientAuthenticatorTest extends TestCase
 {
-    /**
-     * @var CacheInterface|MockInterface
-     */
-    public $cache;
-
     /**
      * @var LwaServiceInterface|MockInterface
      */
@@ -32,25 +25,23 @@ class ClientAuthenticatorTest extends TestCase
     public function setUp()
     {
         parent::setup();
-        $this->cache       = \Mockery::mock(CacheInterface::class);
         $this->lwaService  = \Mockery::mock(LwaServiceInterface::class);
         $this->spApiConfig = $this->buildSpApiConfig();
     }
 
-    public function test_createAuthenticatedGuzzleClient_with_cached_LWA_token()
+    public function test_createAuthenticatedGuzzleClient_with_no_RDT()
     {
-        $cachedLwaToken     = 'fake-lwa-token123';
+        $expectedToken      = 'fake-lwa-token123';
         $credentialProvider = function () {
             return \Mockery::mock(CredentialsInterface::class);
         };
 
-        $this->cache->shouldReceive('get')
+        $this->lwaService->shouldReceive('rememberLwaAccessToken')
             ->once()
-            ->with(ClientAuthenticator::LWA_ACCESS_TOKEN_CACHE_KEY)
-            ->andReturn($cachedLwaToken);
+            ->withNoArgs()
+            ->andReturn($expectedToken);
 
         $sut = new ClientAuthenticator(
-            $this->cache,
             $this->lwaService,
             $credentialProvider,
             $this->spApiConfig
@@ -60,75 +51,28 @@ class ClientAuthenticatorTest extends TestCase
         $guzzleConfig = $guzzleClient->getConfig();
 
         $this->assertFalse(empty($guzzleConfig['headers']['x-amz-access-token']));
-        $this->assertEquals($cachedLwaToken, $guzzleClient->getConfig()['headers']['x-amz-access-token']);
-        $this->assertEquals($this->spApiConfig->spApiBaseUrl, $guzzleClient->getConfig()['base_uri']);
-    }
-
-    public function test_createAuthenticatedGuzzleClient_with_newly_requested_LWA_token()
-    {
-        $newLwaToken        = 'fake-lwa-token123';
-        $expiresIn          = 3600;
-        $credentialProvider = function () {
-            return \Mockery::mock(CredentialsInterface::class);
-        };
-
-        $this->cache->shouldReceive('get')
-            ->once()
-            ->with(ClientAuthenticator::LWA_ACCESS_TOKEN_CACHE_KEY)
-            ->andReturn(null);
-        $this->lwaService->shouldReceive('requestNewLwaAccessToken')
-            ->once()
-            ->withArgs(function ($arg1) {
-                return $arg1 instanceof Client
-                    && $arg1->getConfig('base_uri') == $this->spApiConfig->lwaOAuthBaseUrl;
-            })
-            ->andReturn([
-                'access_token' => $newLwaToken,
-                'expires_in'   => $expiresIn,
-            ]);
-        $this->cache->shouldReceive('set')
-            ->once()
-            ->with(
-                ClientAuthenticator::LWA_ACCESS_TOKEN_CACHE_KEY,
-                $newLwaToken,
-                $expiresIn - ClientAuthenticator::CACHE_LIFE_BUFFER_IN_SECONDS
-            )
-            ->andReturn(true);
-
-        $sut = new ClientAuthenticator(
-            $this->cache,
-            $this->lwaService,
-            $credentialProvider,
-            $this->spApiConfig
-        );
-
-        $guzzleClient = $sut->createAuthenticatedGuzzleClient();
-        $guzzleConfig = $guzzleClient->getConfig();
-
-        $this->assertFalse(empty($guzzleConfig['headers']['x-amz-access-token']));
-        $this->assertEquals($newLwaToken, $guzzleClient->getConfig()['headers']['x-amz-access-token']);
+        $this->assertEquals($expectedToken, $guzzleClient->getConfig()['headers']['x-amz-access-token']);
         $this->assertEquals($this->spApiConfig->spApiBaseUrl, $guzzleClient->getConfig()['base_uri']);
     }
 
     public function test_createAuthenticatedGuzzleClient_with_RDT()
     {
-        $rdt                = 'fake-rdt123';
+        $expectedRdt        = 'fake-rdt123';
         $credentialProvider = function () {
             return \Mockery::mock(CredentialsInterface::class);
         };
 
         $sut = new ClientAuthenticator(
-            $this->cache,
             $this->lwaService,
             $credentialProvider,
             $this->spApiConfig
         );
 
-        $guzzleClient = $sut->createAuthenticatedGuzzleClient($rdt);
+        $guzzleClient = $sut->createAuthenticatedGuzzleClient($expectedRdt);
         $guzzleConfig = $guzzleClient->getConfig();
 
         $this->assertFalse(empty($guzzleConfig['headers']['x-amz-access-token']));
-        $this->assertEquals($rdt, $guzzleClient->getConfig()['headers']['x-amz-access-token']);
+        $this->assertEquals($expectedRdt, $guzzleClient->getConfig()['headers']['x-amz-access-token']);
         $this->assertEquals($this->spApiConfig->spApiBaseUrl, $guzzleClient->getConfig()['base_uri']);
     }
 }
